+++ tests/Builder/Pdf/LibreOfficePdfBuilderTest.php
@@ -8,7 +8,6 @@ use Sensiolabs\GotenbergBundle\Builder\Pdf\LibreOfficePdfBuilder;
 use Sensiolabs\GotenbergBundle\Enumeration\SplitMode;
 use Sensiolabs\GotenbergBundle\Exception\InvalidBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Exception\MissingRequiredFieldException;
-use Sensiolabs\GotenbergBundle\Formatter\AssetBaseDirFormatter;
 use Sensiolabs\GotenbergBundle\Test\Builder\GotenbergBuilderTestCase;
 use Sensiolabs\GotenbergBundle\Tests\Builder\Behaviors\LibreOfficeTestCaseTrait;
 use Symfony\Component\DependencyInjection\Container;
@@ -96,23 +95,5 @@ class LibreOfficePdfBuilderTest extends GotenbergBuilderTestCase
             ->splitUnify()
             ->generate()
         ;
-    }
-
-    public function testWithFileToEmbed(): void
-    {
-        $this->withGotenbergVersion('8.25.0');
-        $this->container->set('asset_base_dir_formatter', new AssetBaseDirFormatter(self::FIXTURE_DIR, [self::FIXTURE_DIR]));
-
-        $this->getBuilder()
-            ->files('assets/office/document.odt')
-            ->filename('testEmbedLibreOffice.pdf')
-            ->embeds('embed/facturX.xml')
-            ->generate()
-        ;
-
-        $this->assertGotenbergFormDataFile('files', 'application/xml', self::FIXTURE_DIR.'/embed/facturX.xml');
-
-        $this->assertGotenbergEndpoint('/forms/libreoffice/convert');
-        $this->assertGotenbergHeader('Gotenberg-Output-Filename', 'testEmbedLibreOffice.pdf');
     }
 }
