+++ tests/Builder/Pdf/MergePdfBuilderTest.php
@@ -6,9 +6,9 @@ use Sensiolabs\GotenbergBundle\Builder\BuilderInterface;
 use Sensiolabs\GotenbergBundle\Builder\Pdf\MergePdfBuilder;
 use Sensiolabs\GotenbergBundle\Exception\InvalidBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Exception\MissingRequiredFieldException;
-use Sensiolabs\GotenbergBundle\Formatter\AssetBaseDirFormatter;
 use Sensiolabs\GotenbergBundle\Test\Builder\GotenbergBuilderTestCase;
 use Sensiolabs\GotenbergBundle\Tests\Builder\Behaviors\DownloadFromTestCaseTrait;
+use Sensiolabs\GotenbergBundle\Tests\Builder\Behaviors\EmbedTestCaseTrait;
 use Sensiolabs\GotenbergBundle\Tests\Builder\Behaviors\FlattenTestCaseTrait;
 use Sensiolabs\GotenbergBundle\Tests\Builder\Behaviors\MetadataTestCaseTrait;
 use Sensiolabs\GotenbergBundle\Tests\Builder\Behaviors\PdfFormatTestCaseTrait;
@@ -22,6 +22,9 @@ final class MergePdfBuilderTest extends GotenbergBuilderTestCase
 {
     /** @use DownloadFromTestCaseTrait<MergePdfBuilder> */
     use DownloadFromTestCaseTrait;
+
+    /** @use EmbedTestCaseTrait<MergePdfBuilder> */
+    use EmbedTestCaseTrait;
 
     /** @use FlattenTestCaseTrait<MergePdfBuilder> */
     use FlattenTestCaseTrait;
@@ -100,23 +103,5 @@ final class MergePdfBuilderTest extends GotenbergBuilderTestCase
         $this->getBuilder()
             ->generate()
         ;
-    }
-
-    public function testWithFileToEmbed(): void
-    {
-        $this->withGotenbergVersion('8.25.0');
-        $this->container->set('asset_base_dir_formatter', new AssetBaseDirFormatter(self::FIXTURE_DIR, [self::FIXTURE_DIR]));
-
-        $this->getBuilder()
-            ->files('pdf/simple_pdf_1.pdf')
-            ->filename('testEmbedMerge.pdf')
-            ->embeds('embed/facturX.xml')
-            ->generate()
-        ;
-
-        $this->assertGotenbergFormDataFile('files', 'application/xml', self::FIXTURE_DIR.'/embed/facturX.xml');
-
-        $this->assertGotenbergEndpoint('/forms/pdfengines/merge');
-        $this->assertGotenbergHeader('Gotenberg-Output-Filename', 'testEmbedMerge.pdf');
     }
 }
